# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "summary": "Create order",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Idempotency key to prevent order duplication (valid for 24 hours)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order returned from idempotency cache"
          },
          "201": {
            "description": "Order successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Recipient or product not found"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "OrdersController_findAll",
        "summary": "Get orders list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Количество элементов на странице",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Смещение для пагинации",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Поисковый запрос",
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Поле для сортировки",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Направление сортировки: ASC или DESC",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by order status",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minAmount",
            "required": false,
            "in": "query",
            "description": "Minimum order total in cents",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxAmount",
            "required": false,
            "in": "query",
            "description": "Maximum order total in cents",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orders list retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedOrderResponse"
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/batch-approve": {
      "post": {
        "operationId": "OrdersController_batchApprove",
        "summary": "Batch approve orders",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchOrderIdsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Orders approved"
          },
          "400": {
            "description": "Some orders cannot be approved"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/batch-delete": {
      "post": {
        "operationId": "OrdersController_batchDelete",
        "summary": "Batch delete orders",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchOrderIdsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Orders deleted"
          },
          "400": {
            "description": "Some orders cannot be deleted"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "summary": "Get order by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "OrdersController_update",
        "summary": "Update order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Cannot update order"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "OrdersController_remove",
        "summary": "Delete order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order successfully deleted"
          },
          "400": {
            "description": "Cannot delete order"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/{id}/confirm": {
      "patch": {
        "operationId": "OrdersController_confirm",
        "summary": "Confirm order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order successfully confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Cannot confirm order"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/{id}/cancel": {
      "patch": {
        "operationId": "OrdersController_cancel",
        "summary": "Cancel order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order successfully cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Cannot cancel order"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/dashboard/daily-revenue": {
      "get": {
        "operationId": "OrdersController_getDailyRevenue",
        "summary": "Get daily revenue/turnover for sparkline",
        "parameters": [
          {
            "name": "days",
            "required": false,
            "in": "query",
            "description": "Number of days (default 7)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily revenue data retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyRevenueDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/dashboard/status-summary": {
      "get": {
        "operationId": "OrdersController_getOrderStatusSummary",
        "summary": "Get order status summary (draft/confirmed/cancelled counts)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Order status summary retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderStatusSummaryDto"
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/dashboard/revenue-by-products": {
      "get": {
        "operationId": "OrdersController_getRevenueByProducts",
        "summary": "Get revenue by products",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue by products retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RevenueByProductDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/dashboard/revenue-by-recipients": {
      "get": {
        "operationId": "OrdersController_getRevenueByRecipients",
        "summary": "Get revenue by recipients",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue by recipients retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RevenueByRecipientDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/dashboard/total-revenue": {
      "get": {
        "operationId": "OrdersController_getTotalRevenue",
        "summary": "Get total revenue",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Total revenue retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotalRevenueDto"
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/dashboard/turnover-by-products": {
      "get": {
        "operationId": "OrdersController_getTurnoverByProducts",
        "summary": "Get turnover by products",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Turnover by products retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TurnoverByProductDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/dashboard/turnover-by-recipients": {
      "get": {
        "operationId": "OrdersController_getTurnoverByRecipients",
        "summary": "Get turnover by recipients",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Turnover by recipients retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TurnoverByRecipientDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/dashboard/total-turnover": {
      "get": {
        "operationId": "OrdersController_getTotalTurnover",
        "summary": "Get total turnover",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for filtering (ISO 8601 string)",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Total turnover retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotalTurnoverDto"
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "ReceiptMaster API",
    "description": "Multi-tenant SaaS API for managing products, recipients, orders, receipts, and analytics.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "recipientId": {
            "type": "string",
            "description": "ID получателя",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "items": {
            "description": "Позиции заказа",
            "example": [
              {
                "productId": "123e4567-e89b-12d3-a456-426614174000",
                "qty": 2
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          }
        },
        "required": [
          "recipientId",
          "items"
        ]
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "recipient_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "confirmed",
              "cancelled"
            ]
          },
          "subtotal_cents": {
            "type": "number"
          },
          "total_cents": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "created_by": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          },
          "is_locked": {
            "type": "boolean"
          },
          "recipient": {
            "$ref": "#/components/schemas/Recipient"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "receipts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Receipt"
            }
          }
        },
        "required": [
          "id",
          "recipient_id",
          "status",
          "subtotal_cents",
          "total_cents",
          "currency",
          "created_by",
          "created_at",
          "updated_at",
          "is_locked",
          "recipient",
          "items",
          "receipts"
        ]
      },
...
      "BatchOrderIdsDto": {
        "type": "object",
        "properties": {
          "orderIds": {
            "description": "Array of order UUIDs",
            "example": [
              "123e4567-e89b-12d3-a456-426614174000"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "orderIds"
        ]
      },
      "PaginatedOrderResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          }
        },
        "required": [
          "total",
          "offset",
          "limit",
          "data"
        ]
      },
...
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "recipientId": {
            "type": "string",
            "description": "Recipient ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "items": {
            "description": "Order items",
            "example": [
              {
                "productId": "123e4567-e89b-12d3-a456-426614174000",
                "qty": 2
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateOrderItemDto"
            }
          }
        }
      },
...
    }
  }
}